pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "pulmonary-anomaly-detection"
version = "1.0.0"
description = "Unsupervised anomaly detection in chest radiographs using a convolutional autoencoder"
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.24.0",
    "Pillow>=9.5.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "scikit-learn>=1.2.0",
    "scikit-image>=0.20.0",
    "flask>=2.3.0",
    "tqdm>=4.65.0",
    "kaggle>=1.5.13",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "app*"]
